Changed the name of the main file
